PLANNER-1835 Move classes to testutil package or remake to inner ones
