+++ optaplanner-core/src/test/java/org/optaplanner/core/config/solver/testutil/calculator/TestdataCorruptedDifferentValuesCalculator.java
@@ -1,4 +1,4 @@
-package org.optaplanner.core.impl.testdata.score.director;
+package org.optaplanner.core.config.solver.testutil.calculator;
 
 public class TestdataCorruptedDifferentValuesCalculator extends AbstractTestdataDifferentValuesCalculator {
